+++ cmd/main-fu/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "backend/parcer/src/parcer"
+
+func main() {
+	parcer.ScrapingOZON()
+}
